+++ bootstrapper/src/main/java/ch/adv/ui/bootstrapper/Bootstrapper.java
@@ -1,8 +1,8 @@
 package ch.adv.ui.bootstrapper;
 
-import ch.adv.ui.ADVApplication;
+import ch.adv.ui.app.ADVApplication;
 import ch.adv.ui.array.ArrayModule;
-import ch.adv.ui.logic.ADVModule;
+import ch.adv.ui.app.ADVModule;
 import ch.adv.ui.logic.ModuleStore;
 import javafx.application.Application;
 import org.slf4j.Logger;
