+++ module-array/src/main/java/ch/adv/ui/array/ArrayModule.java
@@ -2,7 +2,7 @@ package ch.adv.ui.array;
 
 import ch.adv.ui.access.Parser;
 import ch.adv.ui.access.Stringifyer;
-import ch.adv.ui.logic.ADVModule;
+import ch.adv.ui.app.ADVModule;
 import ch.adv.ui.presentation.Layouter;
 
 /**
