+++ ui-core/src/main/java/ch/adv/ui/access/Parser.java
@@ -1,6 +1,6 @@
 package ch.adv.ui.access;
 
-import ch.adv.ui.logic.model.Session;
+import ch.adv.ui.domain.Session;
 
 /**
  * Session parser
