+++ ui-core/src/main/java/ch/adv/ui/access/Stringifyer.java
@@ -1,7 +1,7 @@
 package ch.adv.ui.access;
 
 
-import ch.adv.ui.logic.model.Session;
+import ch.adv.ui.domain.Session;
 
 /**
  * JSON Serializer
