+++ ui-core/src/main/java/ch/adv/ui/presentation/Layouter.java
@@ -1,19 +1,20 @@
 package ch.adv.ui.presentation;
 
-import ch.adv.ui.logic.model.Snapshot;
-import javafx.scene.layout.Pane;
+import ch.adv.ui.domain.Snapshot;
+import ch.adv.ui.presentation.domain.SnapshotWrapper;
 
 /**
  * Builds JavaFX Elements from business objects contained in a snapshot.
  * Places and orients these elements on a JavaFX Pane.
  */
 public interface Layouter {
+
     /**
      * Layouts the snapshot. Builds necessary JavaFX Elements and places them
      * on a pane, which will be displayed in the session view.
      *
      * @param snapshot to be layouted
-     * @return a JavaFX Pane containing the snapshot elements
+     * @return a wrapper containing a layouted JavaFX Pane and the snapshot
      */
-    Pane layout(Snapshot snapshot);
+    SnapshotWrapper layout(Snapshot snapshot);
 }
