--- ui-core/src/test/java/ch/adv/ui/AdvApplicationTest.java
@@ -1,13 +1,0 @@
-package ch.adv.ui;
-
-import org.junit.Test;
-
-import static org.junit.Assert.assertEquals;
-
-public class AdvApplicationTest {
-
-    @Test
-    public void multiplicationOfZeroIntegersShouldReturnZero() {
-        assertEquals(0, 1 * 0);
-    }
-}
