+++ ui-core/src/test/resources/session1.json
@@ -4,20 +4,16 @@
       "snapshotId": 1,
       "snapshotDescription": "1",
       "elements": [
-
       ],
       "relations": [
-
       ]
     },
     {
       "snapshotId": 2,
       "snapshotDescription": "2",
       "elements": [
-
       ],
       "relations": [
-
       ]
     }
   ],
