+++ ui-core/src/test/resources/session2snapshot1.json
@@ -4,10 +4,8 @@
       "snapshotId": 3,
       "snapshotDescription": "1",
       "elements": [
-
       ],
       "relations": [
-
       ]
     }
   ],
