+++ ui-core/src/test/resources/session2snapshot2.json
@@ -4,10 +4,8 @@
       "snapshotId": 4,
       "snapshotDescription": "2",
       "elements": [
-
       ],
       "relations": [
-
       ]
     }
   ],
